requirements.txt
# Core Flask Framework
Flask>=3.0.0
Werkzeug>=3.0.0

# Database & Migration
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.5
alembic>=1.13.0
SQLAlchemy>=2.0.0

# PostgreSQL Drivers
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Migration Progress Bar
tqdm>=4.66.1

# Authentication & Security
Flask-Login>=0.6.3
Flask-JWT-Extended>=4.6.0

# Email & Communication
Flask-Mail>=0.9.1
Flask-CORS>=4.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# AI & Machine Learning
google-generativeai>=0.8.3
openai>=1.3.0

# Image Processing Core
Pillow>=10.1.0  # Compatible with Python 3.13

# Image Analysis Dependencies
numpy>=1.26.0  # Compatible with Python 3.13
filetype>=1.2.0  # Alternative to python-magic for MIME detection

# Image Hashing (with fallback options)
imagehash>=4.3.1
six>=1.16.0  # Required by imagehash

# HTTP Client & APIs
requests>=2.31.0
httpx>=0.25.0  # Modern HTTP client with better async support

# 3D Model Processing
trimesh>=4.0.5
pygltflib>=1.16.1

# Real-time Communication (Core Only)
flask-socketio>=5.3.6
python-socketio>=5.10.0

# API Documentation
flask-restx>=1.3.0

# Data Serialization
marshmallow>=3.20.1
marshmallow-sqlalchemy>=0.29.0

# Background Tasks (Core Only)
celery>=5.3.6

# Database Connection Pooling
sqlalchemy-utils>=0.41.1
psycopg2-pool>=1.1

# Development Tools
watchdog>=3.0.0  # For development auto-reload
